Make it possible to shutdown our server gracefully (not quite)
Use the hyper API to add the mechanism to gracefully shutdown the
server. This commit doesn't actually hook it up to anything though and
hence this commit on its own doesn't add anything. We should hook this
up to UNIX signal handlers for it to be really useful but we add this
already for the end-to-end test case we're about to add in the following
commit.